1、重新设定将原来保存点脉冲改为保存点坐标 2、调整视图大小，按照实际下位机的行程 3、进入任务前刷新坐标 4、去除独立点、中间点是否出胶 5、重新设定轨迹速度范围 6、视图模式支持直接手动修改坐标，视图刷新

+++ app/src/main/java/com/mingseal/data/dao/PointDao.java
@@ -210,10 +210,10 @@ public class PointDao {
 					while (cursor.moveToNext()) {
 						point = new Point(PointType.POINT_NULL);
 						point.setId(cursor.getInt(cursor.getColumnIndex(TablePoint._ID)));
-						point.setX(cursor.getInt(cursor.getColumnIndex(TablePoint.POINT_X)));
-						point.setY(cursor.getInt(cursor.getColumnIndex(TablePoint.POINT_Y)));
-						point.setZ(cursor.getInt(cursor.getColumnIndex(TablePoint.POINT_Z)));
-						point.setU(cursor.getInt(cursor.getColumnIndex(TablePoint.POINT_U)));
+						point.setX(cursor.getFloat(cursor.getColumnIndex(TablePoint.POINT_X)));
+						point.setY(cursor.getFloat(cursor.getColumnIndex(TablePoint.POINT_Y)));
+						point.setZ(cursor.getFloat(cursor.getColumnIndex(TablePoint.POINT_Z)));
+						point.setU(cursor.getFloat(cursor.getColumnIndex(TablePoint.POINT_U)));
 						PointParam pointParam = new PointParam();
 						pointParam.set_id(cursor.getInt(cursor.getColumnIndex(TablePoint.POINT_PARAM_ID)));
 						pointParam.setPointType(PointType.valueOf(PointType.class,
